% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cameca_stat_X.R
\docType{data}
\name{cameca_stat_X}
\alias{cameca_stat_X}
\title{Reference Cameca X statistics}
\format{
A data frame with 54.000 rows and 11 variables:
\describe{
  \item{file.nm}{name of the original file}
  \item{correction block}{systematic corrections (named blocks in Cameca
  output)}
  \item{Mass#}{ion mass according to number of detection device}
  \item{Cumulated count}{Total ion count (same as N~tot~)}
}
}
\usage{
cameca_stat_X
}
\description{
The single ion count statistics as performed by the Cameca software. The ion
counts have not been corrected for systematic offsets by EM detection
devices
}
\keyword{datasets}
