% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{cor_IC}
\alias{cor_IC}
\title{Process raw ion count data}
\usage{
cor_IC(df, N, t, Det, deadtime = 44, thr_PHD = 50)
}
\arguments{
\item{df}{A tibble containing raw ion count data}

\item{N}{A variable constituting the ion counts}

\item{t}{A variable constituting the time increments}

\item{Det}{A character string or variable for the detection system ("EM" or
"FC")}

\item{deadtime}{A numeric value for the deadtime of the EM system}

\item{thr_PHD}{A numeric value for the disrcriminator threshold of the EM
system}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing processed
ion count data and metadata
}
\description{
\code{cor_IC} function for processing ion count data
}
\details{
The accuracy of pulsed ion counting is influenced by systematic errors which
depending on the ion counting system. Dead time and EM yield are two
prominent effects for the electron multiplier systems. The dead time where
the system does not register counts; this occurs when to incident ions strike
the EM in a small enough time window in which the EM channel is
electronically paralysed. The EM yield is the ratio between the number of
output pulses counted after the EM  discriminator threshold and the number of
ions arriving at the EM. The latter can be gauged with the peak height
distribution (PHD) which is the probability for an EM output to have a
certain voltage amplitude.
}
\examples{
# Use point_example() to access the examples bundled with this package in the
# inst/extdata directory.

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

}
