% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{cor_yield}
\alias{cor_yield}
\alias{cor_DT}
\title{Correct ion detection bias}
\usage{
cor_yield(x = NULL, mean_PHD, SD_PHD, thr_PHD, output = "ct")

cor_DT(x, deadtime)
}
\arguments{
\item{x}{A numeric vector containing raw ion count data.}

\item{mean_PHD}{A numeric vector containing the mean PHD value.}

\item{SD_PHD}{A numeric vector containing the standard deviation of the
PHD.}

\item{thr_PHD}{A numeric value for the disrcriminator threshold of the EM
system.}

\item{output}{Character string indicating whether to return corrected count
rates (\code{"ct"}) or yield value (\code{"Y"}).}

\item{deadtime}{A numeric value for the dead-time of the EM system
with units nanoseconds.}
}
\value{
A numeric vector with the corrected count rates.
}
\description{
\code{cor_yield} function to correct counting bias related to EM Yield.
\code{cor_DT} function to correct counting bias related to deadtime
}
\details{
The accuracy of pulsed ion counting is influenced by systematic errors which
depend on the ion counting system. Deadtime and EM yield are two
prominent effects for the electron multiplier systems. The deadtime refer to
the timewindow when the system does not register counts; this occurs when
incident ions strike the EM in a small enough time window in which the EM
channel is electronically paralysed. The EM yield is the ratio between the
number of output pulses counted after the EM  discriminator threshold and
the number of ions arriving at the EM. The latter can be gauged with the peak
height distribution (PHD) which is the probability for an EM output to have a
certain voltage amplitude.
}
\examples{
# Original count rate of a chemical species
x <- 30000

# Corrected count rate for EM Yield with a threshold of 50 V
cor_yield(x, mean_PHD = 210, SD_PHD = 60, thr_PHD = 50)

# Corrected count rate for a deadtime of 44 ns
cor_DT(x, 44)
}
