% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{cov_R}
\alias{cov_R}
\title{Function for co-variate conversion of isotope systems}
\usage{
cov_R(.IC, .ion, ..., .species = species.nm, .t = t.nm, .preserve = FALSE)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{.ion}{A character string or vector constituting ion names.}

\item{...}{Variables for grouping.}

\item{.species}{A variable constituting the species analysed.}

\item{.t}{A variable constituting the time of the analyses.}

\item{.preserve}{A logical whether to preserve ID variable.}
}
\value{
A \code{tibble::\link[tibble:tibble]{tibble}} in wide format
}
\description{
\code{cov_R} create a wide format tibble for an isotope pair
}
\details{
This functions converts the long format data frame to a wide format tibble
for an isotope pair based on an unique identifier for the time of
measurement. The data can be linked by a combination of three variable unique
for the analyses; the file name, the chemical species name and the time
increment of measurement. Pay attention when using in combination with
`zeroCt()`.
}
\examples{

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# wide format
cov_R(tb_pr, c("13C", "12C"), file.nm)

}
