% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{des_SD_prop}
\alias{des_SD_prop}
\alias{hat_SD_prop}
\title{Propagation of error for isotope ratios}
\usage{
des_SD_prop(M_R, ion1, ion2, M_ion1, M_ion2, S_ion1, S_ion2, n, type = "sd")

hat_SD_prop(M_R, ion1, ion2, N_ion1, N_ion2, n, type = "sd")
}
\arguments{
\item{M_R}{A numeric vector constituting the mean isotope ratio}

\item{ion1}{A numeric vector constituting the single ion count rate of the
heavy isotope}

\item{ion2}{A numeric vector constituting the single ion count rate of the
light isotope}

\item{M_ion1}{A numeric vector constituting the mean ion count of the heavy
isotope}

\item{M_ion2}{A numeric vector constituting the mean ion count of the light
isotope}

\item{S_ion1}{A numeric vector constituting the standard deviation of the ion
counts of the heavy isotope}

\item{S_ion2}{A numeric vector constituting the standard deviation of the ion
counts of the light isotope}

\item{n}{A numeric vector constituting the length of the count block \emph{n}}

\item{type}{A character string for the type of uncertainty estimate:
\code{"sd"}, standard deviation; \code{"rsd"}, relative standard deviation in
per mille; \code{"se"}, standard error of the mean; \code{"rse"}, relative
standard error of the mean in per mille}

\item{N_ion1}{A numeric vector constituting the single ion counts of the
heavy isotope}

\item{N_ion2}{A numeric vector constituting the single ion counts of the
light isotope}
}
\value{
A numeric vector containing the propagated uncertainty of the isotope
ratio
}
\description{
\code{des_SD_prop} function for propagation of descriptive error statistics
for isotope ratios (R). \code{hat_SD_prop} function for propagation of
predictive (Poisson) error statistics for isotope ratios (R).
}
\details{
Isotope ratios are based on two measured variables, i.e., two isotopes of a
single chemical species. The combined variable compounds the uncertainty
associated with the individual measurement. Hence error propagation is
required to obtain a reliable estimate of the uncertainty associated
with the compounded variable (i.e. isotope ratio).
}
\examples{
# light isotope count rates
`32S` <- c(22318.19, 22151.20, 22429.52, 22143.78, 22574.25, 22455.50,
         22516.73, 22414.68, 22288.50, 22327.47)

# heavy isotope count rate
`34S` <- c(231.4838, 220.3725, 255.5584, 237.0395, 244.4471, 238.8914,
         238.8914, 237.0395, 262.9660, 264.8179)

# mean 32S
M_32S <- mean(`32S`)

# mean 34S
M_34S <- mean(`34S`)

# mean R
M_R <- M_34S / M_32S

# standard deviation 32S
S_32S <- sd(`32S`)

# standard deviation 34S
S_34S <- sd(`34S`)

# propagation
des_SD_prop(M_R, `34S`, `32S`, M_34S, M_32S, S_34S, S_32S, 10, "rse")

}
