% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{expr_R}
\alias{expr_R}
\title{Create stat_R call quosure}
\usage{
expr_R(Xt, N, species, ion1, ion2)
}
\arguments{
\item{Xt}{A character string constituting the ion count rate.}

\item{N}{A character string constituting the ion counts.}

\item{species}{A character string constituting the species analysed.}

\item{ion1}{A character string constituting the heavy isotope ("13C").}

\item{ion2}{A character string constituting the light isotope ("12C").}

\item{t}{A character string constituting the time increment of measurement.}
}
\value{
A list containing the input arguments as a
\code{\link[rlang:quosure]{quosure}}.
}
\description{
\code{expr_R} and \code{expr_cor} functions generate an quosure that mimics a
stat_R call or cor_IC call for subsequent usage in dia_R
}
\details{
The \code{diag_R} function performs an internal call to stat_R or cor_IC call
to perform diagnostics on the influence of individial measurements on the
blockwise or global (i.e., a complete analysis) stastics. This function
provides a convenient way to enter this call as an quosure into the argument
\code{args} of \code{diag_R}.
}
\examples{
expr_R(Xt = "Xt.pr", N = "N.pr", species = "species.nm", ion1 = "13C",
       ion2 = "12C")
}
