% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-parser.R
\name{formula_parser}
\alias{formula_parser}
\title{Linear model parser}
\usage{
formula_parser(
  data,
  arg1,
  arg2,
  flag = NULL,
  transformation = NULL,
  nest = NULL,
  type = "OLS",
  execute = TRUE
)
}
\arguments{
\item{data}{ion count data.}

\item{arg1}{quosure of outcome variable.}

\item{arg2}{quosure of predictor variable.}

\item{flag}{quosure of nominal predictor variable for outliers.}

\item{transformation}{transformation of the predictor ("log" or "ppt"
dividing through thousand).}

\item{nest}{quosure of nest variable for mixed effect moddeling this is the
variable for the nesting structure.}

\item{type}{type of model fitting (ratio method (RM), ordinary least squares
(OLS), generalized least squares (GLS), special case for Quasi Simultaneous
Arrival estimate (QSA) or linear mixed effects (LME).}

\item{execute}{either evaluate or return expression.}
}
\value{
call or evaluated call in context of data.
}
\description{
Linear model parser
}
