% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{ICdir_chk}
\alias{ICdir_chk}
\title{Check if directory is suitable for point}
\usage{
ICdir_chk(directory, types = c("is_txt", "chk_is", "stat"))
}
\arguments{
\item{directory}{A path or connection to a directory containing raw ion count
data files.}

\item{types}{Regular expression for the required file extensions. Default
searches for files ending with .is_txt, .chk_is, and .stat}
}
\value{
A logical indicating whether the directory is suitable for `point`
}
\description{
This function checks whether the necessary files for the `point` read
functions are included in the directory.
}
\examples{
ICdir_chk(point_example("2018-01-19-GLENDON"))
}
