% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{latex_parser}
\alias{latex_parser}
\title{Latex parsable chemical species names}
\usage{
latex_parser(ion1, ion2 = NULL)
}
\arguments{
\item{ion1}{A chemical species character string (for isotope ratios the heavy
isotope).}

\item{ion2}{A chemical species character string (for isotope ratios the light
isotope).}
}
\value{
A character string parsable in Latex
}
\description{
\code{latex_parser} Converts a character string containing chemical species
names in a latex parsable string.
}
\details{
This functions converts chemical species names of the form, e.g. `"12C"`,
`"13C2"`, or `"12C 14N"`, to a character string which can be parsed in Latex
to species names with appropiate superscripts on th left for mass and
subscripts for stochiometry on the right. If both arguments are given; e.g.,
`ion1 = "13C"` and `ion2 = "12C"`, the function will create an isotope ratio
seperate.
}
