% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{mlm_RS}
\alias{mlm_RS}
\title{Relative standard deviation of the coefficient}
\usage{
mlm_RS(sum, arg, output = "value", trans = TRUE)
}
\description{
Relative standard deviation of the coefficient
}
