% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_diag.R
\docType{data}
\name{names_diag}
\alias{names_diag}
\title{The plot types for diagnostic tests available in \code{point}}
\format{
A data frame with 2 rows and 3 variables:
\describe{
  \item{argument}{argument names in point}
  \item{point}{variable names generated by point::diag_R()}
  \item{use}{usage in point package}
}
}
\source{
\code{point::\link[point:diag_R]{diag_R}()}
}
\usage{
names_diag
}
\description{
A dataset containing the diagnostic point variable names
}
\keyword{datasets}
