% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_model.R
\docType{data}
\name{names_model}
\alias{names_model}
\title{The regression model names associated with
\code{point::\link[point:diag_R]{diag_R}()}}
\format{
A data frame with 7 rows and 4 variables:
\describe{
  \item{name}{statistic name}
  \item{derived}{variable name}
  \item{type}{type of curve fitting}
  \item{label}{description of the statistic}
}
}
\source{
\code{point::\link[point:diag_R]{diag_R}()}
}
\usage{
names_model
}
\description{
A dataset containing the specifications for producing variable names and
latex parsable labels for printing in Rmarkdown as well as details of the
statistics output as and the procedure for finding the best fitting curve.
}
\keyword{datasets}
