% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_plot.R
\docType{data}
\name{names_plot}
\alias{names_plot}
\title{The plot types for diagnostic tests available in \code{point}}
\format{
A data frame with 9 rows and 5 variables:
\describe{
  \item{name}{name of diagnostic test}
  \item{inference}{type of inference test: package internal or external}
  \item{label}{plot title}
  \item{xaxis}{plot xaxis title}
  \item{yaxis}{plot yaxis title}
}
}
\source{
\code{point::\link[point:diag_R]{diag_R}()}
}
\usage{
names_plot
}
\description{
A dataset containing the names and attributes of plots for diagnostic test in
\code{point}
}
\keyword{datasets}
