% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_stat_R.R
\docType{data}
\name{names_stat_R}
\alias{names_stat_R}
\title{The statistics associated with
\code{point::\link[point:stat_R]{stat_R}()}}
\format{
A data frame with 10 rows and 3 variables:
\describe{
  \item{name}{statistic name}
  \item{derived}{variable name}
  \item{origin}{variable name before transformation (ratio)}
  \item{label}{description of the statistic}
}
}
\source{
\code{point::\link[point:stat_R]{stat_R}()}
}
\usage{
names_stat_R
}
\description{
A dataset containing the specifications for producing variable names and
latex parsable labels for printing in Rmarkdown as well as details of the
statistics available in \code{point::\link[point:stat_R]{stat_R}()}.
}
\keyword{datasets}
