% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_stat_X.R
\docType{data}
\name{names_stat_X}
\alias{names_stat_X}
\title{The statistics associated with
\code{point::\link[point:stat_X]{stat_X}()}}
\format{
A data frame with 10 rows and 3 variables:
\describe{
  \item{name}{statistic name}
  \item{derived}{variable name}
  \item{label}{description of the statistic}
}
}
\source{
\code{point::\link[point:stat_X]{stat_X}()}
}
\usage{
names_stat_X
}
\description{
A dataset containing the specifications for producing variable names and
latex parsable labels for printing in Rmarkdown as well as details of the
statistics available in \code{point::\link[point:stat_X]{stat_X}()}.
}
\keyword{datasets}
