% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_IC.R
\name{plot_diag_R}
\alias{plot_diag_R}
\title{Plotting diagnostics}
\usage{
plot_diag_R(
  df,
  Xt,
  N,
  species,
  ion1,
  ion2,
  ...,
  path = NULL,
  device = "png",
  aug = TRUE,
  width.out = 20,
  height.out = 20
)
}
\arguments{
\item{df}{A tibble containing processed ion count data.}

\item{...}{Variables for grouping.}

\item{method}{Character string for the type of diagnostics. Currently only
"standard" is supported, which pertains to the default Cameca software
setting.}

\item{args}{A list of quosures pertaining to the variables required for a
call to stat_R. The function expr_R can be used to simplify setting this
argument.}
}
\value{
A t\code{\link[ggplot2:ggplot]{ggplot}}
}
\description{
\code{plot_RDiag} function for propagation of uncertainty for single ions
}
\details{
The \code{plot_RDiag} visualises the diagnostics
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

# plotting Cook's D diagnostics
plot_RDiag(tb.pr, Xt.pr, N.pr, species.nm, "13C", "12C", file.nm)


}
