\name{POINT-package}
\alias{POINT-package}
\docType{package}
\title{
Protein Structure Guided Local Test
}
\description{
A rare variant association test that utilizes protein tertiary structure to increase signal and to identify likely causal variants. Performs structure-guided collapsing, which leads to local tests that borrow information from neighboring variants on a protein and that provide association information on a variant-specific level.}
\details{
\tabular{ll}{
Package: \tab POINT\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-03-02\cr
License: \tab GPL-2\cr
Depends:  \tab methods, stats, rARPACK, Matrix\cr
Imports: \tab  CompQuadForm
}
}
\author{
Rachel Marceau West, Shannon T. Holloway \cr
Maintainer: Shannon T. Holloway <sthollow@ncsu.edu>
}
\references{
Marceau West R, Lu W, Rotroff DM, Kuenemann MA, Chang SM, et al. (2019) Identifying individual risk rare variants using protein structure guided local tests (POINT). PLOS Computational Biology 15(2): e1006722.
}
