% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{point_example}
\alias{point_example}
\title{Get path to point example}
\usage{
point_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
This function comes from the package `readr`, and has been modified to access
the bundled datatsets in directory `inst/extdata` of `point`. This
function make them easy to access. This function is modified from
\code{\link[readr:readr_example]{readr_example}} of the package
\code{\link[readr]{readr}}.
}
\examples{
point_example()
point_example("2018-01-19-GLENDON")
}
