% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{predict_var}
\alias{predict_var}
\title{Predicting variation in ion based on ionization trend}
\usage{
predict_var(df, Xt, N, t, species, ion, ...)
}
\description{
Predicting variation in ion based on ionization trend
}
