% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{QSA_test}
\alias{QSA_test}
\title{QSA test}
\usage{
QSA_test(.df, .Xt, .N, .species, .ion1, .ion2, ..., .plot = TRUE)
}
\arguments{
\item{df}{A tibble containing raw ion count data.}

\item{N}{A variable constituting the ion counts.}

\item{t}{A variable constituting the time increments.}

\item{Det}{A character string or variable for the detection system ("EM" or
"FC").}

\item{deadtime}{A numeric value for the deadtime of the EM system.}

\item{thr_PHD}{A numeric value for the disrcriminator threshold of the EM.
system}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing the original dataset
and adds the variables: \code{Xt.rw}, ion count rates uncorrected for
detection device-specific biases; \code{Xt.pr}, ion count rates corrected for
detection device-specific biases; and \code{N.pr}, counts corrected for
detection device-specific biases.
}
\description{
\code{QSA_test} function for processing ion count data.
}
\details{
The accuracy of pulsed ion counting is influenced by systematic errors which
depend on the ion counting system. Deadtime and EM yield are two
prominent effects for the electron multiplier systems. The deadtime refer to
the timewindow when the system does not register counts; this occurs when
incident ions strike the EM in a small enough time window in which the EM
channel is electronically paralysed. The EM yield is the ratio between the
number of output pulses counted after the EM  discriminator threshold and
the number of ions arriving at the EM. The latter can be gauged with the peak
height distribution (PHD) which is the probability for an EM output to have a
certain voltage amplitude.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

# QSA test
tb.QSA <- QSA_test(tb.pr, Xt.pr, N.pr, species.nm, "13C", "12C", file.nm)
}
