% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_R.R
\docType{data}
\name{reference_R}
\alias{reference_R}
\title{The reference materials used for common isotope systems after the
\href{https://wwwrcamnl.wr.usgs.gov/isoig/isopubs/itchch2.html}{USGS})}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{isotope}{name of the rare isotope}
  \item{ratio}{isotope ratio}
  \item{reference}{name of reference material}
  \item{value}{value of refrence material}
}
}
\source{
<https://wwwrcamnl.wr.usgs.gov/isoig/isopubs/itchch2.html>
}
\usage{
reference_R
}
\description{
A dataset containing the reference materials and their assigned values for
calibrating isotope systems. These values are used by
\code{point::\link[point:calib_R]{calib_R}()}.
}
\keyword{datasets}
