% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_IC.R
\name{sim_R}
\alias{sim_R}
\title{Simulate ion count data}
\usage{
sim_R(
  n = 3000,
  N_range = 10^6,
  reps = 1,
  ion1,
  ion2,
  sys,
  type,
  baseR = NULL,
  offsetR = NULL,
  seed,
  ...
)
}
\arguments{
\item{sys}{Systematic component as relative standard deviation of the major/minor single ion in per mille}
}
\description{
Simulate ion count data
}
