% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_IC.R
\name{simu_R}
\alias{simu_R}
\title{Simulate ion count data}
\usage{
simu_R(
  .sys,
  .type,
  .ion1,
  .ion2,
  .reference,
  .seed,
  .n = 3000,
  .N = 1e+06,
  .bl = 50,
  .reps = 1,
  .baseR = 0,
  .devR = 0,
  ...
)
}
\arguments{
\item{.sys}{Systematic variation caused by ionization fluctuations as
relative standard deviation of the major ion in per mille}

\item{.type}{Character string to select the type of simulation
\code{"symmetric"}, \code{"asymmetric"} and \code{"ideal"}, where the former
two types introduce an offset caused by a deviation in R.}

\item{.ion1}{A character string constituting the heavy isotope ("13C").}

\item{.ion2}{A character string constituting the light isotope ("12C").}

\item{.reference}{Character string for conversion of delta values to R
(.e.g. VPDB; see \code{?calib_R()} for more information).}

\item{.seed}{Numeric sees for reproducibility of the generated data.}

\item{.n}{Numeric for the number of measurements.}

\item{.N}{Numeric for total ion count of the light isotope.}

\item{.bl}{Numeric for block number.}

\item{.reps}{Multiplication of the procedure (e.g. effectively generating
multiple analyses).}

\item{.baseR}{Numeric for the baseline isotope value in delta notation in per
mille.}

\item{.devR}{Numeric for the deviation (or forcing) away from the baseline as
delta notation in per mille.}

\item{...}{Not supported currently}
}
\value{
Tibble with simulated ion count data.D
}
\description{
Simulate ion count data
}
\examples{

# Gradient in 13C/12C over measurement transect
simu_R(50, "symmetric", "13C", "12C", "VPDB", 1)

}
