% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{stat_R}
\alias{stat_R}
\title{function for propagation of uncertainty in ion ratios (isotope values)}
\usage{
stat_R(df, Xt, N, ID, ion1, ion2, ..., latex = FALSE)
}
\description{
function for propagation of uncertainty in ion ratios (isotope values)
}
