% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{stat_SDprop}
\alias{stat_SDprop}
\title{Propagation of errors for isotope ratios}
\usage{
stat_SDprop(ion1, ion2, type = "sd", predicted = FALSE)
}
\arguments{
\item{ion1}{A numeric vector constituting the single ion count rate of the
heavy isotope if argument predicted is set to \code{FALSE}, otherwise counts
are required.}

\item{ion2}{A numeric vector constituting the single ion count rate of the
light isotope if argument predicted is set to \code{FALSE}, otherwise counts
are required.}

\item{type}{A character string for the type of uncertainty estimate:
\code{"sd"}, standard deviation; \code{"rsd"}, relative standard deviation in
per mille; \code{"se"}, standard error of the mean; \code{"rse"}, relative
standard error of the mean in per mille.}

\item{predicted}{Logical indicating whether sd is descriptive \code{FALSE},
or based on predicted Poisson value \code{TRUE}.}
}
\value{
A numeric vector containing the propagated uncertainty of the isotope
ratio.
}
\description{
\code{stat_SDprop} function for propagation of descriptive and predictive
(Poisson) error statistics for isotope ratios (R).
}
\details{
Isotope ratios are based on two measured variables, i.e., two isotopes of a
single chemical species. The combined variable compounds the uncertainty
associated with the individual measurement. Hence error propagation is
required to obtain a reliable estimate of the uncertainty associated
with the compounded variable (i.e. isotope ratio).
}
\examples{
# Light isotope count rates
`32S` <- c(22318.19, 22151.20, 22429.52, 22143.78, 22574.25, 22455.50)

# Heavy isotope count rate
`34S` <- c(231.4838, 220.3725, 255.5584, 237.0395, 244.4471, 238.8914)

# Propagation
stat_SDprop(`34S`, `32S`, "rse")
}
