% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{stat_X}
\alias{stat_X}
\alias{stat_R}
\title{Analyse raw ion count data}
\usage{
stat_X(
  .IC,
  ...,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .stat = point::names_stat_X$name,
  .label = NULL,
  .output = "sum"
)

stat_R(
  .IC,
  .ion1,
  .ion2,
  ...,
  .nest = NULL,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .stat = point::names_stat_R$name,
  .label = NULL,
  .output = "sum",
  .zero = FALSE
)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{...}{Variables for grouping.}

\item{.X}{A variable constituting the ion count rate (defaults to
variables generated with \code{read_IC()}.)}

\item{.N}{A variable constituting the ion counts (defaults to variables
generated with \code{read_IC()}.).}

\item{.species}{A variable constituting the species analysed (defaults to
variables generated with \code{read_IC()}.).}

\item{.t}{A variable constituting the time of the analyses (defaults to
variables generated with \code{read_IC()}.).}

\item{.stat}{Select statistics (e.g. \code{c("M", "RS"}), see the tables
\code{point::names_stat_X} and \code{point::names_stat_R} for the full
selection of statistics available (default uses all statistic
transformations).}

\item{.label}{A character string indicating whether variable names are latex
(\code{"latex"}) or webtex (\code{"webtex"}) compatible. Will be extended in
the future \code{default = NULL}.}

\item{.output}{A character string for output as summary statistics ("sum");
statistics only ("stat"); and statistics with the original data ("complete")
\code{default = "sum"}.}

\item{.ion1}{A character string constituting the heavy isotope ("13C").}

\item{.ion2}{A character string constituting the light isotope ("12C").}

\item{.nest}{A variable hat identifies a series of analyses to calculate
external precision.}

\item{.zero}{A character string that determines whether analyses with zero
count measurements will be removed from the calculations.}
}
\value{
A \code{tibble::\link[tibble:tibble]{tibble}} containing descriptive
and predictive statistics for ion counts and isotope ratios. The naming
convention depends on the argument \code{latex}; if set to \code{FALSE},
variable names concerning statistics will consist of an abbreviation pasted
together with the input variable names of \code{Xt}.
}
\description{
\code{stat_X} function for descriptive and predictive statistics on single
ion precision.
\code{stat_R} function for descriptive and predictive statistics on isotope
ratios (R) precision with appropriate error propagation.
}
\details{
These functions are a convenient wrapper to calculate the statistics
pertaining to the precision of pulsed ion count data (e.g. secondary ion mass
spectrometry). The statistics can either be calculated for single ions or
isotope ratios, and include observed and predicted (Poisson) statistics.
Calculations for isotope ratios include proper error propagation. For more
information on the usage as well as the mathematics behind these
functions see \code{vignette("IC-precision", package = "point")}.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# Single ion descriptive an predictive statistics for all measured ions
stat_X(tb_pr, file.nm)

# Descriptive an predictive statistics for 13C/12C ratios
stat_R(tb_pr, "13C", "12C", file.nm, .zero = TRUE)

# Descriptive an predictive statistics for 13C/12C ratios (external)
stat_R(tb_pr, "13C", "12C", sample.nm, file.nm, .nest = file.nm,
       .zero = TRUE)
}
