% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twodens.R
\name{twodens}
\alias{twodens}
\title{Plotting highly dens data (inspired by
\href{https://themockup.blog/posts/2020-08-28-heatmaps-in-ggplot2/}{Thomas Mock})}
\usage{
twodens(.IC, .x, .y, ..., .flag)
}
\arguments{
\item{.IC}{A tibble containing ion count data.}

\item{.x}{The variable plotted on the x-axis.}

\item{.y}{The variable plotted on the y-axis.}

\item{...}{Variables for grouping.}

\item{.flag}{Variable identifying two components in the data.}
}
\value{
A
\code{tibble::\link[tibble:tibble]{tibble}()} with the 2D density of x and y
named \code{dens_z}.
}
\description{
\code{twodens} function for 2D density calculations of one or two components
}
\details{
This function is inspired by code in the blog-post by
\href{https://themockup.blog/posts/2020-08-28-heatmaps-in-ggplot2/}{Thomas Mock},
which deals with over-plotting of very dens data. This is just the problem
that occurs when dealing with SIMS ion count data, which usually entails
thousands of data-points (measurements) for just one analysis. The function
provided here can deal with density of one or two components, where in the
latter case the difference of two 2D density matrices is used to show the
area on the plot with a maximum difference.
}
