% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{unfold}
\alias{unfold}
\alias{fold}
\title{Access and hide IC metadata}
\usage{
unfold(df, type = "metadata", merge = TRUE)

fold(df, type, meta = NULL)
}
\arguments{
\item{df}{A tibble containing ion count data along any point of the point-
workflow}

\item{type}{A character string identifying the metadata (default:
\code{"metadata"})}

\item{merge}{Logical dictating whether metadata is joined to the tibble or
returned as a separate file.}

\item{meta}{Additional tibble containing the metadata for storage along the
main IC data.}
}
\value{
A tibble with metadata as an attribute, columns or as a seperate
tibble.
}
\description{
\code{unfold()} helps unpack metadata associated with ion count
data loaded with \code{read_IC()}. \code{fold()} does the opposite an hides
the metadata as attribute of the tibble.
}
\examples{
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"), hide = TRUE)

# Unfold metadata
unfold(tb_rw, merge = FALSE)
}
