% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{zeroCt}
\alias{zeroCt}
\title{Remove analytical runs with zero counts}
\usage{
zeroCt(.IC, .ion1, .ion2, ..., .N = NULL, .species = NULL, .warn = TRUE)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{.ion1}{A character string constituting the rare isotope ("13C").}

\item{.ion2}{A character string constituting the common isotope ("12C").}

\item{...}{Variables for grouping.}

\item{.N}{A variable constituting the ion counts.}

\item{.species}{A variable constituting the species analysed.}

\item{.warn}{A logical indicating whether to produce a warning.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing the single ion count
dataset for the specified ion ratio. The grouping variable specifies on
which level the original dataset is subsetted.
}
\description{
\code{zeroCt} removes analytical runs for isotope ratios that contain zero
counts.
}
\details{
This functions removes analytical runs with zero counts for calculating
isotope ratios.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# Remove analyses with zero counts
tb_pr <- zeroCt(tb_pr, "13C", "12C", file.nm)
}
