% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_IC.R
\name{calib_R}
\alias{calib_R}
\title{Calibration of isotope ratios}
\usage{
calib_R(
  x,
  y = NULL,
  standard,
  type = c("composition", "enrichment"),
  input = c("R", "F", "delta", "delta_alpha"),
  output = c("R", "F", "delta", "alpha")
)
}
\arguments{
\item{x}{A numeric value or vector.}

\item{y}{A numeric value or vector of product when calculating enrichments (`type = enrichment`).}

\item{standard}{A character or numeric value or vector.}

\item{type}{Type of conversion, isotope composition or enrichment.}

\item{input}{The type of input value, R, F or delta.}

\item{output}{Desired ouput value, R, F or delta.}
}
\value{
A numeric value or vector.
}
\description{
\code{calib_R} function to convert isotope values back-and-forth between R and delta
}
\details{
A fundamental in publishing isotope data is the conversion of isotope the delta
formulation comparing the obtained R to that of a standard. These values are
reported on a per mill scale. However in IC calculations most often we use
ratios (R) or fractional abundances (F). This functions provide an easy way
of making these transformaitons for both the average composition but also
fractionations as enrichment factors.
}
\examples{
# R value of 0.0111 to delta on VPDB scale
calib_R(0.0111, standard = 0.011237, type = "composition", input = "R", output = "delta")
# or
calib_R(0.0111, standard = "VPDB", type = "composition", input = "R", output = "delta")

# Fractional abundance for a given R
calib_R(0.0111, type = "composition", input = "R", output = "F")

# Generates warning as input and output are equal
calib_R(0.0111, type = "composition", input = "R", output = "R")

# Generates warning as standard is unknown
calib_R(0.0111, standard = "somestandard", type = "composition", input = "R", output = "delta")

# Find R values 60 per mille apart
A <- calib_R(5, standard = "VPDB", type = "composition", input = "delta", output = "R")
B <- calib_R(-55, standard = "VPDB", type = "composition", input = "delta", output = "R")
}
