% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cameca.R, R/CooksD.R
\name{Cameca}
\alias{Cameca}
\alias{CooksD}
\title{Family of diagnostics functions for isotope count ratios}
\usage{
Cameca(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = Xt.pr,
  .N = N.pr,
  .t = t.nm,
  .output = "complete",
  .alpha_level = 0.05,
  .hyp = "none"
)

CooksD(
  .IC,
  .ion1,
  .ion2,
  ...,
  .X = Xt.pr,
  .t = t.nm,
  .output = "complete",
  .hyp = "none",
  .alpha_level = 0.05
)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{.ion1}{A character string constituting the heavy isotope ("13C").}

\item{.ion2}{A character string constituting the light isotope ("12C").}

\item{...}{Variables for grouping.}

\item{.X}{A variable constituting the ion count rate (defaults to
variables generated with \code{read_IC()})}

\item{.N}{A variable constituting the ion counts (defaults to variables
generated with \code{read_IC()}.).}

\item{.t}{A variable constituting the time of the analyses (defaults to
variables generated with \code{read_IC()}.).}

\item{.output}{Character string determining whether the returned values in a
minimal version `"flag"` (original dataset + diagnostics) or an extended
version with all the intermediate steps of ion- and isotope-wise summary
statistics `"complete"`.}

\item{.alpha_level}{The significance level of the hypothesis test and
rejection level for outliers.}

\item{.hyp}{Hypothesis test appropriate for the selected method.}
}
\value{
A \code{tibble\link[tibble:tibble]{tibble}()} containing either the
original dataset with new columns related to the diagnostics or only the
diagnostics. The flag variable enables convenient filtering of the original
tibble for an augmentation of the original dataset.
}
\description{
\code{Cameca} default CAMECA diagnostics
\code{CooksD} regression diagnostics based on Cook's D
}
\details{
These functions perform a specific set of diagnostics to term anomalous
values in raw ion count data of an isotope pair. The wrapper function
\code{diag_R} is more convenient as it defines all the ion- and isotope-wise
statistics required for the diagnostics.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# Descriptive an predictive statistics for 13C/12C ratios (note .output
# argument and remove zero count analysis)
tb_R <- stat_R(tb_pr, "13C", "12C", file.nm, sample.nm, .output = "complete",
               .zero = TRUE)

# CAMECA style augmentation of ion count data for isotope ratios
Cameca(tb_R, "13C", "12C", file.nm, .output = "flag")
}
