% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{cor_IC}
\alias{cor_IC}
\title{Process raw ion count data}
\usage{
cor_IC(df, N, t, Det, deadtime = 44, thr = 180)
}
\arguments{
\item{df}{A tibble containing raw ion count data}

\item{N}{A variable constituting the ion counts}

\item{deadtime}{A numeric value for the deadtime of the EM}

\item{Xt}{A variable constituting the ion count rate}

\item{...}{Variables for grouping}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing processed
ion count data and metadata
}
\description{
\code{proc_Xt} function for processing ion count data
}
\details{
Ion count data consisting of time-incremented integer values are process. These functions
are currently only supported for data generated by a NanoSIMS50L. Raw ion
count data and accompayning is extracted and collated into a single tible
from text files with the extensions \emph{.is_txt} and \emph{.stat},
respectively. These files can be found in the directories associated with the
SIMS measurements.
}
\examples{
# Use point_example() to access the examples bundled with this package in the
# inst/extdata directory. The examples directories are named:
# 2020-01-17-TREASURE and "2018-01-19-GLENDON"

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

}
