% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{diag_R}
\alias{diag_R}
\title{Diagnostics isotope count data}
\usage{
diag_R(
  df,
  method = "Cameca",
  args = expr_R(NULL),
  reps = 1,
  ...,
  plot = TRUE,
  plot_type = "interactive"
)
}
\arguments{
\item{df}{A tibble containing processed ion count data.}

\item{method}{Character string for the type of diagnostics. "Cameca" pertains
to the default Cameca software setting. "CooksD" pertains to global
regression diagnostics based on Cook's D statistics.}

\item{args}{A list of quosures pertaining to the variables required for a
call to stat_R. The function expr_R can be used to simplify setting this
argument.}

\item{reps}{The number of iterations of diagnostic-augmentation cycles.}

\item{...}{Variables for grouping.}

\item{plot}{Logical indicating whether plot is produced, if FALSE a list of
tibbles with length `reps` is produced.}

\item{Character}{string determing whether the returned a \code{"static"} plot
or an \code{"interactive"} plot with \code{\link[plotly:ggplotly]{ggplotly}}.}
}
\value{
A list containing several \code{\link[tibble:tibble]{tibble}} the
elements named df contain the original dataset and augmented versions from
element number 2 onward. The elements named results contain the flag
variable for outlier detection as well as associated statistics of the
selected procedure.
}
\description{
\code{diag_R} wrapper function for diagnostics on isotope count data
}
\details{
The \code{diag_R} function performs an internal call to stat_R to perform
diagnostics on the influence of individial measurements on the blockwise or
global (i.e., a complete analysis) stastics. It identifies potentially
influential measurements that can subsequentally be filtered from the
original dataset to improve the precision. See
\code{vignette("IC-diagnostics", package = "point")} for more information
on how to use the function, and possible methods for diagnostics.
}
\examples{
# Modelled ion count dataset
# Create expression for mimicking R_stat() call
expr_R_stat <- expr_R(Xt = "Xt.sim",
                      N = "N.sim",
                      species = "species",
                      ion1 = "13C",
                      ion2 = "12C"
                      )

# Cook's D style diagnostic-augmentation of ion count data for
# isotope ratios; 3 repeats
diag_R(sim_IC_extremes,
       method = "CooksD",
       args = expr_R_stat,
       reps = 3,
       simulation,
       trend
       )
}
