% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{diag_R}
\alias{diag_R}
\title{Diagnostics isotope count data}
\usage{
diag_R(df, method = "Cameca", args = expr_R(NULL), ..., output = "flag")
}
\arguments{
\item{df}{A tibble containing processed ion count data.}

\item{method}{Character string for the type of diagnostics. "Cameca" pertains
to the default Cameca software setting. "CooksD" pertains to global
regression diagnostics based on Cook's D statistics.}

\item{args}{A list of quosures pertaining to the variables required for a
call to stat_R. The function expr_R can be used to simplify setting this
argument.}

\item{...}{Variables for grouping.}

\item{output}{Character string determing whether the returned values in a
minimal version `"flag"` (original dataset + diagnostics) or an extended
version with all the intermediate steps of ion- and isotope-wise summary
statistics `"complete"`.}
}
\value{
A t\code{\link[tibble:tibble]{tibble}} containing either the original
dataset with new columns related to the diagnostics an extended version with
all the intermediate steps of ion- and isotope-wise summary statistics. The
flag variable enables convenient filtering of the original tibble for an
augmentation of the original dataset.
}
\description{
\code{diag_R} wrapper function for diagnostics on isotope count data
}
\details{
The \code{diag_R} function performs an internal call to stat_R to perform
diagnostics on the influence of individial measurements on the blockwise or
global (i.e., a complete analysis) stastics. It identifies potentially
influential measurements that can subsequentally be filtered from the
original dataset to improve the precision. See
\code{vignette("IC-diagnostics", package = "point")} for more information
on how to use the function, and possible methods for diagnostics.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

# CAMECA style augmentatio of ion count data for isotope ratios
tb.aug <- diag_R(tb.pr,
                 method = "Cameca",
                 args = expr_R(Xt = "Xt.pr",
                               N = "N.pr",
                               species = "species.nm",
                               ion1 = "13C",
                               ion2 = "12C"),
                 file.nm,
                 bl.mt)
}
