% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{diag_R}
\alias{diag_R}
\title{Diagnostics raw ion count data}
\usage{
diag_R(df, method = "standard", Moreargs = quos(NULL), ...)
}
\arguments{
\item{df}{A tibble containing processed ion count data}

\item{method}{Character string for the type of diagnostics}
}
\value{
A tibble containing diagnostics
}
\description{
\code{diag_R} function for propagation of uncertainty for single ions
}
\details{
Ion count data consisting of time-incremented integer values are
}
\examples{
# Use system.file() to access the examples bundled with this package in the
# inst/extdata directory. The examples directories are named:
# 2020-01-17-TREASURE and "2018-01-19-GLENDON"


tb.dia <- diag_R(tb.pr, method = "standard", Moreargs = quos(Xt = Xt.pr,
                 N = N.pr, species = species.nm, ion1 = "13C", ion2 = "12C",
                 .named = TRUE), file.nm)

}
