% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{diag_R}
\alias{diag_R}
\title{Diagnostics isotope count data}
\usage{
diag_R(
  .IC,
  .ion1,
  .ion2,
  ...,
  .nest = NULL,
  .method = "CooksD",
  .reps = 1,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .output = "inference",
  .label = "none",
  .meta = FALSE,
  .alpha_level = 0.05,
  .hyp = "none",
  .plot = FALSE,
  .plot_type = "static",
  .plot_stat = NULL,
  .plot_iso = FALSE,
  .plot_outlier_labs = c("divergent", "confluent")
)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{.ion1}{A character string constituting the rare isotope ("13C").}

\item{.ion2}{A character string constituting the common isotope ("12C").}

\item{...}{Variables for grouping.}

\item{.nest}{A variable hat identifies a series of analyses to calculate
the significance of inter-isotope variability.}

\item{.method}{Character string for the method for diagnostics (default =
\code{"CooksD"}, see details).}

\item{.reps}{Numeric setting the number of repeated iterations of outlier
detection (default = 1).}

\item{.X}{A variable constituting the ion count rate (defaults to
variables generated with \code{read_IC()})}

\item{.N}{A variable constituting the ion counts (defaults to variables
generated with \code{read_IC()}.).}

\item{.species}{A variable constituting the species analysed (defaults to
variables generated with \code{read_IC()}).}

\item{.t}{A variable constituting the time of the analyses (defaults to
variables generated with \code{read_IC()}).}

\item{.output}{Can be set to \code{"complete"} which returns \code{stat_R()}
and \code{stat_X()} statistics, diagnostics, and inference test results
following the selected method (see above argument \code{.method});
\code{"augmented"} for the augmented IC data after diagnostics;
\code{"diagnostic"} returns \code{stat_R()} and \code{stat_X()} statistics
and outlier detection; \code{"outlier"} for outlier detection;
\code{"inference"} for only inference test statistics results
(default = \code{"inference"}).}

\item{.label}{For printing nice latex labels use \code{"latex"} (default =
\code{NULL}).}

\item{.meta}{Logical whether to preserve the metadata as an attribute
(defaults to TRUE).}

\item{.alpha_level}{Significance level of hypothesis test.}

\item{.hyp}{Hypothesis test appropriate for the selected method
(default = \code{"none"}).}

\item{.plot}{Logical indicating whether plot is generated.}

\item{.plot_type}{Character string determining whether the returned plot is
\code{"static"} \code{ggplot2::\link[ggplot2:ggplot2]{ggplot2}()}(currently
only supported option).}

\item{.plot_stat}{Adds a statistic label to the plot (e.g. . \code{"M"}), see
\code{point::nm_stat_R} for the full selection of statistics available.}

\item{.plot_iso}{A character string (e.g. \code{"VPDB"}) for the delta
conversion of R (see \code{?calib_R()} for options).}

\item{.plot_outlier_labs}{A character vector of length two for the colourbar
text for outliers (default = c("divergent", "confluent")).}
}
\value{
A \code{ggplot2::\link[ggplot2:ggplot]{ggplot}()} is returned
(if \code{.plot = TRUE}) along with a
\code{tibble::\link[tibble:tibble]{tibble}()} which can contain statistics
diagnostics, hypothesis test results associated with the chosen method and
depending on the argument \code{.output}.
}
\description{
\code{diag_R} wrapper function for diagnostics on isotope count data
}
\details{
The \code{diag_R} function performs an internal call to stat_R to perform
diagnostics on the influence of individual measurements on the block-wise or
global (i.e., a complete analysis) statistics. It identifies potentially
influential measurements that indicate heterogeneity in the analytical
substrate or measurement. See
\code{vignette("IC-diagnostics", package = "point")} for more information on
how to use the function, and possible methods for outlier detection. Options
are \code{"CooksD"} (default), \code{"Cameca"}, \code{"Rm"}, \code{"norm_E"},
\code{"CV"}, \code{"IR"}, and \code{"QQ"}, see the
\code{vignette("IC-diagnostics", package = "point")} for examples and
\code{point::\link[point:names_plot]{names_plot}}. The
argument \code{.output} can be used to toggle between \code{"complete"};
returning \code{stat_R()} and \code{stat_X()} statistics, diagnostics, and
inference test results, \code{"augmented"}; returning the augmented IC after
removing outliers, \code{"diagnostic"}; for only outlier detection results;
\code{"diagnostic"}; for statistics and outlier detection, or
\code{"inference"}; returns only inference test statistics results
(default = inference).
}
\examples{
# Modelled ion count dataset
# Cook's D style diagnostic-augmentation of ion count data for
# isotope ratios
diag_R(simu_IC, "13C", "12C", type.nm, spot.nm, .plot = TRUE)

}
