% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_diagnostics.R
\name{eval_diag}
\alias{eval_diag}
\title{Evaluate effect size and significance of outliers on R}
\usage{
eval_diag(
  .IC,
  .ion1,
  .ion2,
  ...,
  .nest = NULL,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .flag = NULL,
  .execution = NULL,
  .output = "inference",
  .tf = "ppt",
  .label = NULL,
  .meta = TRUE
)
}
\arguments{
\item{.IC}{A tibble containing ion count data and diagnostics generated with
\code{diag_R()}, as a minimum the outlier \code{flag} variable is required.}

\item{.ion1}{A character string constituting the rare isotope (e.g. "13C").}

\item{.ion2}{A character string constituting the common isotope (e.g. "12C").}

\item{...}{Variables for grouping.}

\item{.nest}{A variable hat identifies a series of analyses to calculate
the significance of inter-isotope variability.}

\item{.X}{A variable constituting the ion count rate (defaults to
variables generated with \code{read_IC()})}

\item{.N}{A variable constituting the ion counts (defaults to variables
generated with \code{read_IC()}.).}

\item{.species}{A variable constituting the species analysed (defaults to
variables generated with \code{read_IC()}).}

\item{.t}{A variable constituting the time of the analyses (defaults to
variables generated with \code{read_IC()}).}

\item{.flag}{A variable constituting the outlier flag (defaults to
variables generated with \code{diag_R()}).}

\item{.execution}{A variable constituting the iterative cycles of diagnostics
(defaults to variables generated with \code{diag_R()}).}

\item{.output}{A character string for output as summary statistics
("inference") and statistics with the original data ("complete").}

\item{.tf}{Variable transformation as parts per thousand (\code{"ppt"}) or
log (\code{"log"}) before mixed linear model application.}

\item{.label}{A character string indicating whether variable names are latex
(\code{"latex"}) or webtex (\code{"webtex"}) compatible. Will be extended in
the future \code{default = NULL}.}

\item{.meta}{Logical whether to preserve the metadata as an attribute
(defaults to TRUE).}
}
\value{
A \code{tibble::\link[tibble:tibble]{tibble}()} with model output.
See \code{point::names_model} for more information on the model results.
}
\description{
\code{eval_diag} function for the evaluation of effect size and significance
of outliers on R detected with diagnostics, such as Cook's D or sigma
rejection (Cameca default method).
}
\examples{
# Simulated IC data
tb_dia <- diag_R(simu_IC, "13C", "12C", type.nm, spot.nm,
                 .output = "diagnostic")

# Evaluate significance and effect of outliers based on Cook's D
eval_diag(tb_dia, "13C", "12C", type.nm, spot.nm, .nest = type.nm)

}
