% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_IC.R
\name{eval_diag}
\alias{eval_diag}
\title{Evaluate effect size of diagnostics}
\usage{
eval_diag(ls_df, args, flag, ..., nest = FALSE, group = NULL, output = "sum")
}
\arguments{
\item{ls_df}{A list of tibbles as generated by \code{diag_R}.}

\item{args}{A list of quosures pertaining to the variables required for a
call to stat_R. The function expr_R can be used to simplify setting this
argument.}

\item{...}{Variables for grouping.}

\item{output}{A character string for output as summary statistics ("sum") and
statistics with the original data ("complete").}

\item{df2}{A dataframe with the ionization tend of the major ion in per
mille.}

\item{RSXt}{A variable constituting the relative standard deviation of the
major ion in the dataframe enetered for the argument \code{df2}.}

\item{n}{A variable indicating the number of observation in the list of
dataframes entered for the argument \code{ls_df}}
}
\description{
\code{eval_diag} function for the evaluation of effect size of diagnostics
}
\examples{
# evaluation of diagnostics
tb.eval <- eval_diag(ls.dia,
                     args = expr_R_stat,
                     flag = flag,
                     simulation,
                     trend
                     )
}
