% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_cameca.R
\docType{data}
\name{names_cameca}
\alias{names_cameca}
\title{The data associated with Cameca NanoSIMS 50L}
\format{
A data frame with 48 rows and 5 variables:
\describe{
  \item{point}{variable names generated by point::\link[point:read_IC]{read_IC}()}
  \item{argument}{argument names in point}
  \item{use}{usage in point package}
  \item{cameca}{names of Cameca NanoSIMS 50L meta-data during point measurement}
  \item{extension}{Cameca file extension}
  \item{format}{format of data in Cameca file}
  \item{definition}{definitions of variables}
}
}
\source{
\code{point::\link[point:read_IC]{read_IC}()}
}
\usage{
names_cameca
}
\description{
A dataset containing the original and point variable names of Cameca NanoSIMS
50L meta-data used by \code{point::\link[point:read_IC]{read_IC}()}.
}
\keyword{datasets}
