% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\docType{package}
\name{point}
\alias{point}
\title{point: A package for reading, processing, and analysing raw ion count data.}
\description{
The point package provides function for reading processing and analysing
pulsed ion count data.
}
\details{
The point functions provide a convenient way to work with and evaluate pulsed
ion count data. The \code{read_*} functions can load and collate raw ion
count data in a \code{tibble::\link[tibble:tibble]{tibble}()}. Systematic
biases that are introduced by the ion detection devices can be corrected with
the \code{cor_*} functions. The precision of ion count data can be checked
with the \code{stat_*} functions. The \code{diag_*} functions provide a
convenient way to measure the influence of individual measurements
(\emph{N}) on an \emph{n}-series of measurements (or analysis).
}
