% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\docType{package}
\name{point}
\alias{point}
\title{point: A package for reading, processing, and analysing raw ion count data.}
\description{
The point package provides three categories of important functions: read,
process and analyse that assist with a convenient way of interpreting ion
count data.
}
\section{point functions}{

The point functions ...
}

