% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_trends.R
\name{predict_ionize}
\alias{predict_ionize}
\title{Predicting trends in ionization efficiency}
\usage{
predict_ionize(
  .IC,
  ...,
  .nest = NULL,
  .X = Xt.pr,
  .N = N.pr,
  .species = species.nm,
  .t = t.nm,
  .plot = TRUE,
  .method = "median",
  .hide = TRUE
)
}
\arguments{
\item{.IC}{A tibble containing raw ion count data.}

\item{...}{Variables for grouping.}

\item{.nest}{A variable identifying a groups of analyses which indicates
whether a nested mixed GAM model is applied.}

\item{.X}{A variable constituting the ion count rate.}

\item{.N}{A variable constituting the ion counts.}

\item{.species}{A variable constituting the species analysed.}

\item{.t}{A variable constituting the time increments.}

\item{.plot}{Logical indicating whether to plot ion trends}

\item{.method}{Method for calculating de-trended single ion counts}

\item{.hide}{A logical indicating whether only processed data should be
returned. If \code{TRUE} The model parameters are contained as an attribute
named \code{"modeldata"}.}
}
\description{
Fluctuations in electronics, development of the sputter pit geometry and the
analysed substrate can all cause trends and fluctuations in the secondary ion
current. This function attempts to accommodate the global trend in the
ionization trend by application of a GAM model. The nested variant can then
be used to gauge whether a set of analyses are likely to originate from a
homogeneous substrate at the level of individual analysis.
}
\examples{
# raw ion counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# remove zero count analysis
tb_0 <- zeroCt(tb_pr, "12C", "40Ca 16O", sample.nm, file.nm, .warn = FALSE)

# predict ionization trends
predict_ionize(tb_0, file.nm)
}
