% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_IC.R
\name{predict_ionize}
\alias{predict_ionize}
\title{Predicting trends in ionization efficiency}
\usage{
predict_ionize(df, Xt, N, t, species, ..., .plot = TRUE)
}
\description{
Predicting trends in ionization efficiency
}
