% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSA_test.R
\name{QSA_test}
\alias{QSA_test}
\title{QSA test}
\usage{
QSA_test(
  .IC,
  .ion1,
  .ion2,
  ...,
  .nest = NULL,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL,
  .plot = TRUE
)
}
\arguments{
\item{.IC}{A tibble containing processed ion count data.}

\item{.ion1}{A character string constituting the heavy isotope ("13C").}

\item{.ion2}{A character string constituting the light isotope ("12C").}

\item{...}{Variables for grouping.}

\item{.nest}{A variable hat identifies a series of analyses to calculate
the significance of QSA.}

\item{.X}{A variable constituting the ion count rate (defaults to
variables generated with \code{read_IC()}.)}

\item{.N}{A variable constituting the ion counts (defaults to variables
generated with \code{read_IC()}.).}

\item{.species}{A variable constituting the species analysed (defaults to
variables generated with \code{read_IC()}.).}

\item{.t}{A variable constituting the time of the analyses (defaults to
variables generated with \code{read_IC()}.).}

\item{.plot}{Currently not supported.}
}
\value{
A \code{tibble::\link[tibble:tibble]{tibble}()} containing the
original dataset and adds the variables: \code{beta}, \code{t_QSA}, and
\code{p_QSA} that summarise the results of an linear model fitted by OLS
(respectively; the slope and the associated student's t test statistic and
p value) on the ion count rates of the common isotope (as predictor) and the
isotope ratio (as dependent variable). The p value is for \code{beta} being
different from zero.
.
}
\description{
\code{QSA_test} function to test for QSA.
}
\details{
The accuracy of pulsed ion counting is influenced by systematic errors which
depend on the ion counting system. Quasi simultaneous arrival is one of
those potential errors that can also impact isotope ratios.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb_rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb_pr <- cor_IC(tb_rw)

# QSA test
QSA_test(tb_pr, "13C", "12C", file.nm)
}
