% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{read_IC}
\alias{read_IC}
\alias{read_meta}
\title{Read raw ion count data}
\usage{
read_IC(directory)

read_meta(directory)
}
\arguments{
\item{directory}{A path or connection to a directory containing raw ion count
data txt  files.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing raw ion count data
and metadata
}
\description{
\code{read_IC()} is designed to obtain the numerical data associated with ion
counts.
\code{read_meta()} can be used to specifically retrieve the metadate
associated with ion count data analysis, thereby loading specifications
related to the optics, the primary and secondary ion beams, and the mass
spectrometer.
}
\details{
Ion count data consists of time-incremented integer values. These functions
are currently only supported for data generated by a NanoSIMS50L. Raw ion
count data and accompayning is extracted and collated into a single tibble
from text files with the extensions \emph{.is_txt} \emph{.chk_is} and
\emph{.stat},respectively. These files can be found in the directories
associated with the SIMS measurements.
}
\examples{
# Use point_example() to access the examples bundled with this package

read_IC(point_example("2018-01-19-GLENDON"))
}
