% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw.R
\name{read_IC}
\alias{read_IC}
\alias{read_meta}
\title{Read raw ion count data}
\usage{
read_IC(directory)

read_meta(directory)
}
\arguments{
\item{directory}{A path or connection to a directory containing raw ion count
data txt  files.}
}
\value{
A tibble containing raw ion count data and metadata
}
\description{
\code{read_IC()} is designed to obtain the numerical data associated with ion
counts.
\code{read_meta()} can be used to solely retrieve the metadate associated with ion
count data analysis; loading specifications related to optics, the primary
and secondary ion beam, and mass spectrometer settings.
}
\details{
Ion count data consists of time-incremented integer values. These functions
are currently only supported for data generated by a NanoSIMS50L. Raw ion
count data and accompayning is extracted and collated into a single tible
from text files with the extensions \emph{.is_txt} and \emph{.stat},
respectively. These files can be found in the directories associated with the
SIMS measurements.
}
\section{Functions}{
\itemize{
\item \code{read_meta}: 
}}

\examples{
# Use system.file() to access the examples bundled with this package in the
# inst/extdata directory. The examples directories are named:
# 2020-01-17-TREASURE and "2018-01-19-GLENDON"

read_IC(system.file("extdata", "2020-01-17-TREASURE", package = "point"))#'

}
