% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{stat_Xt}
\alias{stat_Xt}
\alias{stat_R}
\title{Analyse raw ion count data}
\usage{
stat_Xt(df, Xt, N, species, ..., latex = FALSE, output = "sum")

stat_R(
  df,
  Xt,
  N,
  species,
  ion1,
  ion2,
  ...,
  latex = FALSE,
  output = "sum",
  zero = FALSE
)
}
\arguments{
\item{df}{A tibble containing processed ion count data.}

\item{Xt}{A variable constituting the ion count rate.}

\item{N}{A variable constituting the ion counts.}

\item{species}{A variable constituting the species analysed.}

\item{...}{Variables for grouping.}

\item{latex}{A logical indicating whether variable names are latex
compatible.}

\item{output}{A character string for output as summary statistics ("sum");
statistics only ("stat"); and statistics with the original data ("complete").}

\item{ion1}{A character string constituting the heavy isotope ("13C").}

\item{ion2}{A character string constituting the light isotope ("12C").}

\item{zero}{A character string that determines whether analyses with zero
count measurements present will be removed from the calculations.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing descriptive and
predictive statistics for ion counts and isotope ratios. The naming
convention depends on the argument \code{latex}; if set to \code{FALSE},
variable names concerning statistics will consist of an abbreviation pasted
together with the input variable names of \code{Xt}.
}
\description{
\code{stat_Xt} function for descriptive and predictive statistics on single
ion precision.
\code{stat_R} function for descriptive and predictive statistics on isotope
ratios (R) precision with appropiate error propagation.
}
\details{
These functions are a convenient wrapper to calculate the statistics
pertaining to the precision of pulsed ion count data (e.g. secondary ion mass
spectrometry). The statistics can either be calculated for single ions or
isotope ratios, and include observed and predicted (Poisson) statistics.
Calculations for isotope ratios include proper error propagation. For more
information on the usage as well as the mathematics behind these
functions see \code{vignette("IC-precision", package = "point")}.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

# Single ion descriptive an predictive statistics for all measured ions
tb.Xt <- stat_Xt(tb.pr, Xt.pr, N.pr, species.nm, file.nm)

# Descriptive an predictive statistics for 13C/12C ratios
tb.R <- stat_R(tb.pr, Xt.pr, N.pr, species.nm, ion1 = "13C",
               ion2 = "12C", file.nm)
}
