% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{stat_Xt}
\alias{stat_Xt}
\title{Analyse raw ion count data}
\usage{
stat_Xt(df, Xt, N, species, ..., latex = FALSE)
}
\arguments{
\item{df}{A tibble containing processed ion count data}

\item{Xt}{A variable constituting the ion count rate}

\item{N}{A variable constituting the ion counts}

\item{...}{Variables for grouping}

\item{ID}{A character string that uniqually identifies ion pairs for
calculating isotope ratios}

\item{ion1}{A character string constituting the heavy isotope ("13C")}

\item{ion2}{A character string constituting the light isotope ("12C")}
}
\value{
A tibble containing descriptive and predictive statistics for ion
count and isotope ratio data
}
\description{
\code{stat_Xt} function for propagation of uncertainty for single ions
\code{stat_R} function for propagation of uncertainty for isotope ratios (R)
}
\details{
Ion count data consisting of time-incremented integer values are process.
These functions are currently only supported for data generated by a
NanoSIMS50L. Raw ion count data and accompayning is extracted and collated
into a single tible from text files with the extensions \emph{.is_txt} and
\emph{.stat}, respectively. These files can be found in the directories
associated with the SIMS measurements.
}
\examples{
# Use system.file() to access the examples bundled with this package in the
# inst/extdata directory. The examples directories are named:
# 2020-01-17-TREASURE and "2018-01-19-GLENDON"

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(system.file("extdata", "2018-01-19-GLENDON", package = "point"))

# processing raw ion count data
tb.pr <- cor_IC(tb.rw, N.rw, t.rw, det_type.mt)

# single ion descriptive an predictive statistics for all measured ions
tb.Xt <- stat_Xt(tb.pr, Xt.pr, N.pr, file.nm, species.nm)

# descriptive an predictive statistics for 13C/12C ratios
tb.R <- stat_R(tb.pr, Xt.pr, N.pr, ID = "ID", ion1 = "13C", ion2 = "12C",
               file.nm, species.nm)

}
