% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{zeroCt}
\alias{zeroCt}
\title{Remove analytical runs with zero counts}
\usage{
zeroCt(df, N, species, ion1, ion2, ..., warn = TRUE)
}
\arguments{
\item{df}{A tibble containing processed ion count data.}

\item{N}{A variable constituting the ion counts.}

\item{species}{A variable constituting the species analysed.}

\item{ion1}{A character string constituting the heavy isotope ("13C").}

\item{ion2}{A character string constituting the light isotope ("12C").}

\item{...}{Variables for grouping.}

\item{warn}{A logical indicating whether to produce a warning.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing the single ion count
dataset for the specified ion ratio. The grouping variable specifies on
which level the original dataset is subsetted.
}
\description{
\code{zeroCt} removes analytical runs for isotope ratios that contain zero
counts.
}
\details{
This functions removes analytical runs with zero counts for calculating
isotope ratios.
}
\examples{
# Use point_example() to access the examples bundled with this package

# raw data containing 13C and 12C counts on carbonate
tb.rw <- read_IC(point_example("2018-01-19-GLENDON"))

# Remove analyses with zero counts
# Vectors of isotope ratios
ion1 <-  c("13C", "12C 13C", "13C 14N", "12C 14N", "12C")
ion2 <-  c("12C", "12C2", "12C 14N", "40Ca 16O", "40Ca 16O")
tb.pr <- purrr::map2(ion1, ion2,
         ~zeroCt(tb.rw, N.rw, species.nm, .x, .y, file.nm))
}
