% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_IC.R
\name{zeroCt}
\alias{zeroCt}
\title{function to detect and remove zero counts of desired isotope system}
\usage{
zeroCt(df, N, ion1, ion2, ...)
}
\description{
function to detect and remove zero counts of desired isotope system
}
